Factor out SalesforceToBrazeTransformations
