+++ handlers/batch-email-sender/src/main/scala/com/gu/batchemailsender/api/batchemail/SalesforceToBrazeTransformations.scala
@@ -1,0 +1,36 @@
+package com.gu.batchemailsender.api.batchemail
+
+import org.joda.time.DateTime
+import org.joda.time.format.DateTimeFormat
+import scala.util.{Failure, Success, Try}
+
+object SalesforceToBrazeTransformations {
+  def fromSfDateToDisplayDate(sfDate: String): String = {
+    val formattedDate: Try[String] = Try {
+      val asDateTime = DateTime.parse(sfDate, DateTimeFormat.forPattern("yyyy-MM-dd"))
+      asDateTime.toString(DateTimeFormat.forPattern("d MMMM yyyy"))
+    }
+
+    formattedDate match {
+      case Success(date) => date
+      case Failure(_) => sfDate
+    }
+  }
+
+  /**
+   * Salesforce mailingStreet field concatenates on a single line (line1,line), whilst MMA has it over two separate lines
+   */
+  private val sfStreetPattern = """([^,]+),(.*)""".r
+
+  def sfStreetToLine1(in: String): Option[String] =
+    in match {
+      case sfStreetPattern(line1, _) if line1.nonEmpty => Some(line1)
+      case _ => None
+    }
+
+  def sfStreetToLine2(in: String): Option[String] =
+    in match {
+      case sfStreetPattern(_, line2) if line2.nonEmpty => Some(line2)
+      case _ => None
+    }
+}
